fix: Переделай большую часть валидации через аннотации feat: Добавь фреймворк для аннотаций в pom файл fix: Раздели тесты на несколько классов fix: Переделай контроллеры в соответствии с аннотациями, там где не получяется использовать аннотации используй StringUtils для строк
