+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -4,6 +4,10 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.PastOrPresent;
 import java.time.LocalDate;
 
 @Data
@@ -11,11 +15,15 @@ import java.time.LocalDate;
 public class User {
     private int id;
     @EqualsAndHashCode.Exclude
+    @Email
     private String email;
     @EqualsAndHashCode.Exclude
+    @NotBlank
     private String login;
     @EqualsAndHashCode.Exclude
     private String name;
     @EqualsAndHashCode.Exclude
+    @PastOrPresent
+    @NotNull
     private LocalDate birthday;
 }
