+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
-logging.level.ru.yandex.practicum.controller=debug
+logging.level.ru.yandex.practicum.filmorate.controller=debug
 logging.level.org.zalando.logbook=TRACE
